+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- [changed] Admin SDK can now create custom tokens without being initialized
+  with service account credentials. When a service account private key is not
+  available, the SDK uses the remote IAM service to sign JWTs in the cloud.
 - [changed] Updated the typings of the `admin.database.Query.once()`
   method to return a more specific type.
 - [changed] Admin SDK can now read the Firebase/GCP project ID from both